requirements.txt
fastapi
httpx
python-dotenv
uvicorn
jinja2
pydantic
pydantic_settings